% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap_within.R
\name{ap_within_block}
\alias{ap_within_block}
\title{AP within-transformation (column-block version)}
\usage{
ap_within_block(
  M,
  fes,
  w = NULL,
  tol = 1e-08,
  maxiter = 30,
  verbose = FALSE,
  block = 64
)
}
\arguments{
\item{M}{numeric matrix (n x k)}

\item{fes}{list of fixed-effect vectors/factors}

\item{w}{numeric weights or NULL}

\item{tol}{convergence tolerance}

\item{maxiter}{max AP iterations}

\item{verbose}{logical}

\item{block}{integer, number of columns per block}
}
\value{
numeric matrix
}
\description{
AP within-transformation (column-block version)
}
\keyword{internal}
